rename has_virtual_hearing_spec to has_virtual_hearing_examples to fix warning
