+++ spec/models/concerns/has_virtual_hearing_examples.rb
