+++ spec/models/legacy_hearing_spec.rb
@@ -2,7 +2,7 @@
 
 require "support/vacols_database_cleaner"
 require "rails_helper"
-require "models/concerns/has_virtual_hearing_spec"
+require "models/concerns/has_virtual_hearing_examples"
 
 describe LegacyHearing, :all_dbs do
   it_should_behave_like "a model that can have a virtual hearing" do
